developing state-machine

+++ Bank.h
@@ -22,12 +22,12 @@ public:
   	void savePreset();
   	void printPreset(int t_numPreset);
 private:
-	const int m_numBanks{ 5 }; // change if more presets are needed
-	const int m_numPresets{ 5 }; // number of presets per Bank
-	const int m_numLoops{ 10 }; // number of fx loops per preset
-	const int m_numLoopsPerPreset{ m_numPresets*m_numLoops }; // number of fx loops per bank
-	int m_currBankNum;
-	int m_currPresetNum;
+	const uint8_t m_numBanks{ 5 }; // change if more presets are needed
+	const uint8_t m_numPresets{ 5 }; // number of presets per Bank
+	const uint8_t m_numLoops{ 10 }; // number of fx loops per preset
+	const uint8_t m_numLoopsPerPreset{ m_numPresets*m_numLoops }; // number of fx loops per bank
+	uint8_t m_currBankNum;
+	uint8_t m_currPresetNum;
 	struct preset m_currPresets[5];
   	struct preset m_tempPreset;
 	void loadPreset(int t_numBank);
